Assorted API tweaks to make high DPI devices easier to work with and more likely to do the Right Thing(tm) by default
- Add ICanvasResourceCreatorWithDpi interface
- Add DPI properties to CanvasBitmap, CanvasDrawingSession, CanvasImageSource, and CanvasControl
- Plumb DPI through from one object to another in more places than it previous was
- CanvasRenderTarget constructor takes size in dips rather than pixels
- CanvasImageSource constructor takes size in dips rather than pixels
- CanvasImageSource.CreateDrawingSession uses dips

+++ tests/ExampleGallery/Shared/BurningTextExample.xaml.cs
@@ -86,7 +86,7 @@ namespace ExampleGallery
         /// </summary>
         private void SetupText(Size targetSize)
         {
-            textRenderTarget = new CanvasRenderTarget(this.canvas, (int)targetSize.Width, (int)targetSize.Height);
+            textRenderTarget = new CanvasRenderTarget(this.canvas, (float)targetSize.Width, (float)targetSize.Height);
 
             using (var ds = textRenderTarget.CreateDrawingSession())
             {
